+++ data-raw/data_usps.R
@@ -12,14 +12,14 @@ usps_b0 <- read_html("https://pe.usps.com/text/pub28/28apb.htm")
 
 # scrape states table
 usps_state <- usps_b0 %>%
-  html_node(css = "#ep18684") %>%
+  html_element(css = "#ep18684") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   set_names(nm = c("full", "abb")) %>%
   mutate(full = str_to_upper(full))
 
 # scrape armed forces table
 usps_armed <- usps_b0 %>%
-  html_node(css = "#ep19241") %>%
+  html_element(css = "#ep19241") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   set_names(nm = c("full", "abb")) %>%
   mutate(full = str_to_upper(str_remove(full, "([:punct:].*)$")))
@@ -34,7 +34,7 @@ usps_state <- usps_state %>%
 usethis::use_data(usps_state, overwrite = TRUE)
 write_lines(
   x = usps_state,
-  path = "data-raw/usps_state.csv"
+  file = "data-raw/usps_state.csv"
 )
 
 # save state names vector
@@ -42,7 +42,7 @@ valid_name <- usps_state$full
 usethis::use_data(valid_name, overwrite = TRUE)
 write_lines(
   x = valid_name,
-  path = "data-raw/valid_name.csv"
+  file = "data-raw/valid_name.csv"
 )
 
 # save state abbs vector
@@ -50,7 +50,7 @@ valid_abb <- usps_state$abb
 usethis::use_data(valid_abb, overwrite = TRUE)
 write_lines(
   x = valid_abb,
-  path = "data-raw/valid_abb.csv"
+  file = "data-raw/valid_abb.csv"
 )
 
 # street words and abbs ---------------------------------------------------
@@ -61,7 +61,7 @@ usps_c1 <- read_html("https://pe.usps.com/text/pub28/28apc_002.htm")
 
 # scrape states table
 usps_street <- usps_c1 %>%
-  html_node(css = "#ep533076") %>%
+  html_element(css = "#ep533076") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   select(2, 3) %>%
   set_names(nm = c("full", "abb")) %>%
@@ -74,12 +74,12 @@ usps_c2 <- read_html("https://pe.usps.com/text/pub28/28apc_003.htm")
 
 # scrape states table
 usps_unit <- usps_c2 %>%
-  html_node(css = "#ep538257") %>%
+  html_element(css = "#ep538257") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   as_tibble(.name_repair = "unique") %>%
   slice(-26) %>%
   select(1, 3) %>%
-  na_if("") %>%
+  type_convert(na = "") %>%
   drop_na() %>%
   set_names(nm = c("full", "abb")) %>%
   mutate(
@@ -90,7 +90,7 @@ usps_unit <- usps_c2 %>%
 
 # scrape geographic directions
 usps_dirs <- usps_b0 %>%
-  html_node(css = "#ep19168") %>%
+  html_element(css = "#ep19168") %>%
   html_table(fill = TRUE, header = TRUE) %>%
   set_names(nm = c("full", "abb")) %>%
   mutate(full = str_to_upper(full))
@@ -99,13 +99,17 @@ usps_dirs <- usps_b0 %>%
 usps_street <- usps_street %>%
   bind_rows(usps_unit) %>%
   bind_rows(usps_dirs) %>%
+  add_row(
+    full = "PLACE",
+    abb = "PL"
+  ) %>%
   arrange(full)
 
 # save usps street abbs tibble
 usethis::use_data(usps_street, overwrite = TRUE)
 write_lines(
   x = usps_street,
-  path = "data-raw/usps_street.csv"
+  file = "data-raw/usps_street.csv"
 )
 
 # city abbs ---------------------------------------------------------------
@@ -133,5 +137,5 @@ usps_city <- usps_street %>%
 usethis::use_data(usps_city, overwrite = TRUE)
 write_lines(
   x = usps_city,
-  path = "data-raw/usps_city.csv"
+  file = "data-raw/usps_city.csv"
 )